Continue `wget` downloads
By default, resume transfers by starting the download at `N` bytes, where `N` is the current size of the local file. Note that this can be dangerous when you want to overwrite an existing file. In that case, you can use `-O your.file` to force overwriting, or `--continue=no` to disable resuming.